pyproject.toml
[tool.poetry]
name = "ctagtbd"
version = "0.1.0"
description = "CTAG TBD python library and CLI"
authors = ["Michael Jonathan Lee <code.mjlee@gmail.com>"]
license = "GPL 3.0"
readme = "README.md"

packages = [
    { include = "tbd_tools", from = "tools" },
]

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.5"
gitpython = "^3.1.43"
jinja2 = "^3.1.4"
kconfiglib = "^14.1.0"
pyhumps = "^3.8.0"
cxxheaderparser = "^1.4.1"
pydantic = "^2.9.2"

[tool.poetry.scripts]
tbd = 'tbd_tools:tbd_tool'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
